Update suggested model names per TODOs
