+++ server/models/index.js
@@ -2,8 +2,8 @@ const User = require ('./User');
 const Comment = require('./Comment');
 const Likes = require('./Likes');
 
-const MemeCreation = require('./MemeCreation');
+const Meme = require('./Meme');
 const MemeTemplate = require('./MemeTemplate');
 
-module.exports = { User, Comment, Likes, MemeCreation, MemeTemplate };
-// Noteq    
+module.exports = { User, Comment, Likes, Meme, MemeTemplate };
+
